+++ rust/psibase_macros/psibase-macros-lib/src/tests/helpers.rs
@@ -1,0 +1,21 @@
+#[cfg(test)]
+use proc_macro2::TokenStream;
+
+#[cfg(test)]
+pub fn assert_tokens_eq(expected: &TokenStream, actual: &TokenStream) {
+    let expected = expected.to_string();
+    let actual = actual.to_string();
+
+    if expected != actual {
+        println!(
+            "{}",
+            colored_diff::PrettyDifference {
+                expected: &expected,
+                actual: &actual,
+            }
+        );
+        println!("expected: {}", &expected);
+        println!("actual  : {}", &actual);
+        panic!("expected != actual");
+    }
+}
